consolidate logic into TestEFBlockchain
